djot.ast: add custom __pairs metatable to nodes.
This enforces a deterministic order of iteration over keys,
and also forces the "long" forms to be used (e.g. 'children' not 'c').

The order is:

- tag
- text
- other elements in alphabetical order
- children
- references
- footnotes

Removed custom json rendering code that was used to force the
long key names to be used.

Our JSON and HTML output should now be deterministic.