Don't allow openers inside a matched pair of `[..]` ...
followed immediately by `(`, `[`, or `{` to match a closer
outside them, even if the construction doesn't turn out to
be a link/span/image.

This amounts to treating the `[` as "closed" by the `]` even
in a non-link/span/image.  I think that's the most intuitive
behavior.  It means that `[x_y](x_y)` gets parsed as a link
rather than text around emphasis.

Closes #88.