Match links via their [ and ( intermediate openers
Avoid issues with underscores in reference labels or destinations.

Add a link opener on the intermediate opener, '[' or '(',
to clear the link text opener for potential emphasis
and to track the link text opener and closer positions.

Keep the behaviour for unclosed reference or destination
to ignore link text delimiters.
A match for emphasis could eventually be added by default,
this is not the case here.

Explicitly limit link text delimiters to underscore.
Without this guard, any matched delimiter pair could introduce a link.