Fix whitespace in negative charclasses
In vim patterns, `\s` inside a "collection" matches literal s, not
space-or-tab. Buried in `:help collection` is this factoid:

    NOTE: The other backslash codes mentioned above do not work inside
    []!

So before this fix, `_something_` wouldn't start an emphasis at the first
underscore because the pattern for the syntax region required underscore
not-followed-by `s`. This diff replaces `\s` inside square bracket
patterns with `[:blank:]`, which is a character class expression for
"space or tab".